Move code to new ext-footnotes module with extension
