+++ commonmark-ext-footnotes/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module org.commonmark.ext.footnotes {
+    exports org.commonmark.ext.footnotes;
+
+    requires org.commonmark;
+}
