+++ commonmark-ext-footnotes/src/main/java/org/commonmark/ext/footnotes/FootnotesExtension.java
@@ -1,0 +1,24 @@
+package org.commonmark.ext.footnotes;
+
+import org.commonmark.Extension;
+import org.commonmark.ext.footnotes.internal.FootnoteBlockParser;
+import org.commonmark.parser.Parser;
+
+/**
+ * TODO
+ */
+// TODO: HTML rendering and Markdown rendering
+public class FootnotesExtension implements Parser.ParserExtension {
+
+    private FootnotesExtension() {
+    }
+
+    public static Extension create() {
+        return new FootnotesExtension();
+    }
+
+    @Override
+    public void extend(Parser.Builder parserBuilder) {
+        parserBuilder.customBlockParserFactory(new FootnoteBlockParser.Factory());
+    }
+}
