+++ commonmark-ext-footnotes/src/main/javadoc/overview.html
@@ -1,0 +1,6 @@
+<html>
+<body>
+<b>Extension for footnotes using [^1] syntax</b>
+<p>See {@link org.commonmark.ext.footnotes.FootnotesExtension}</p>
+</body>
+</html>
